Add a validate function prop
This will allow a user to pass a function that will be called to validate the text before it is added as a tag. Common use case can be a list of emails, that you want to validate an item is a valid email, but can be useful for many things.