Update to Inline Style
Update in order to Fix Code Splitting Issue with Laravel Mix, Laravael Mix has an issue with Styles when code-splitting, this is a Webpack related Issue, Having Inline Styles creates a situation where the component loads sometimes after compilation and does not most times. Inlining the styles and removing the Style tag from the component, Fixes the issue.